+++ apps/web/app/clients/[id]/layout.tsx
@@ -1,0 +1,14 @@
+import { HeaderClient } from '@repo/ui'
+
+export default function ClientLayout({
+  children,
+}: {
+  children: React.ReactNode
+}): JSX.Element {
+  return (
+    <>
+      <HeaderClient />
+      <div className="bg-white text-black">{children}</div>
+    </>
+  )
+}
