+++ apps/web/app/clients/[id]/page.tsx
@@ -1,3 +1,10 @@
+import { NavBar, NavBarData } from '@repo/ui'
 export default function Client(): JSX.Element {
-  return <></>
+  return (
+    <main>
+      <div className="pb-24 md:pb-0">
+        <NavBar navBarData={NavBarData} />
+      </div>
+    </main>
+  )
 }
