--- apps/web/app/profile/page.tsx
@@ -1,3 +1,0 @@
-export default function Profile(): JSX.Element {
-  return <></>
-}
