--- packages/ui/src/components/Header/index.ts
@@ -1,1 +1,0 @@
-export * from './Header'
