+++ packages/ui/src/components/HeaderHome/index.ts
@@ -1,0 +1,1 @@
+export * from './HeaderHome'
