+++ packages/ui/src/components/index.ts
@@ -3,7 +3,8 @@ export { Button } from './Button'
 export * from './Icon'
 export { Sidebar } from './Sidebar'
 export { ToolCard } from './ToolCard'
-export { Header } from './Header'
+export { HeaderHome } from './HeaderHome'
+export { HeaderClient } from './HeaderClient'
 export { ToolsCarousel } from './ToolsCarousel'
 export { ProfileCard } from './ProfileCard'
 export { ProfilesCarousel } from './ProfileCarousel'
